requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appdirs=1.4.4=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
automat=20.2.0=py_0
bcrypt=3.2.0=py37haf1e3a3_0
brotlipy=0.7.0=py37h9ed2024_1003
ca-certificates=2022.4.26=hecd8cb5_0
certifi=2022.5.18.1=py37hecd8cb5_0
cffi=1.15.0=py37hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
constantly=15.1.0=py37h28b3542_0
cryptography=36.0.0=py37hf6deb26_0
cssselect=1.1.0=pyhd3eb1b0_0
cyrus-sasl=2.1.27=h8a25369_8
hyperlink=21.0.0=pyhd3eb1b0_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
incremental=21.3.0=pyhd3eb1b0_0
itemadapter=0.3.0=pyhd3eb1b0_0
krb5=1.19.2=hcd88c3b_0
libcxx=12.0.0=h2f01273_0
libdb=6.2.32=h0a44026_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.3=hb1e8313_2
libiconv=1.16=hca72f7f_2
libprotobuf=3.11.4=hd9629dc_0
libxml2=2.9.12=hcdb78fc_0
libxslt=1.1.34=h83b36ba_0
lxml=4.8.0=py37h65b224f_0
mysql=5.7.24=h56848d4_0
mysql-connector-c=6.1.11=hccea1a4_1
mysql-connector-python=8.0.18=py37h2de1c92_1
ncurses=6.3=hca72f7f_2
openssl=1.1.1o=hca72f7f_0
parsel=1.6.0=py37hecd8cb5_0
pip=21.2.2=py37hecd8cb5_0
protego=0.1.16=py_0
protobuf=3.11.4=py37h0a44026_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydispatcher=2.0.5=py37hecd8cb5_2
pyhamcrest=2.0.2=pyhd3eb1b0_2
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py37hecd8cb5_0
python=3.7.13=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
queuelib=1.5.0=py37hecd8cb5_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
scrapy=2.4.0=py37_0
service_identity=18.1.0=pyhd3eb1b0_1
setuptools=61.2.0=py37hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h707629a_0
tk=8.6.11=h7bc2e8c_0
twisted=22.2.0=py37hca72f7f_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py37hecd8cb5_0
w3lib=1.21.0=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.12=h4dc903c_2
zope=1.0=py37_1
zope.interface=5.4.0=py37h9ed2024_0
